+++ src/containers.ts
@@ -1,4 +1,5 @@
 export { CircularBuffer } from "./containers/circular-buffer.js";
+export { BlockQueue } from "./containers/block-queue.js";
 export { Deque } from "./containers/deque.js";
 export { PriorityQueue } from "./containers/priority-queue.js";
 export { RBTree } from "./containers/rbtree.js";