+++ src/index.ts
@@ -130,6 +130,7 @@ export {
 
 // Data Structures
 export { CircularBuffer } from "./containers/circular-buffer.js";
+export { BlockQueue } from "./containers/block-queue.js";
 export { Deque } from "./containers/deque.js";
 export { PriorityQueue } from "./containers/priority-queue.js";
 export { RBTree } from "./containers/rbtree.js";